Fixed compatibility with dw2pdf
When exporting a page to pdf with mpdf (via the plugin dw2pdf), the page
stopped being rendered at the hidden blocks.
I don't know exactly why, but it appears that closing divs with </div>
instead of </div\n> fixed this issue.

Those \n were needed to ease the work with the DOM (to avoid the
creation of "empty" nodes on some browser), therefor, the DOM is now
dealt by with jQuery